// client/app/src/generated/dto/video.ts
export interface UploadVideoRequest {
    /**
     * @generated from protobuf field: string title = 1
     */
    title: string;
    /**
     * @generated from protobuf field: bytes file = 2
     */
    file: Uint8Array;
    /**
     * @generated from protobuf field: int32 userId = 3
     */
    userId: number;
    /**
     * @generated from protobuf field: bool enableComment = 4
     */
    enableComment: boolean;
    /**
     * @generated from protobuf field: string visibility = 5
     */
    visibility: string;
    /**
     * @generated from protobuf field: bytes thumbnail = 6
     */
    thumbnail: Uint8Array;
    /**
     * @generated from protobuf field: bool isDraft = 7
     */
    isDraft: boolean;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp postDate = 8
     */
    postDate?: Timestamp;
}
// ...
class UploadVideoRequest$Type extends MessageType<UploadVideoRequest> {
    constructor() {
        super("UploadVideoRequest", [
            { no: 1, name: "title", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "file", kind: "scalar", T: 12 /*ScalarType.BYTES*/ },
            { no: 3, name: "userId", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 4, name: "enableComment", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 5, name: "visibility", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 6, name: "thumbnail", kind: "scalar", T: 12 /*ScalarType.BYTES*/ },
            { no: 7, name: "isDraft", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 8, name: "postDate", kind: "message", T: () => Timestamp }
        ]);
    }
    create(value?: PartialMessage<UploadVideoRequest>): UploadVideoRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.title = "";
        message.file = new Uint8Array(0);
        message.userId = 0;
        message.enableComment = false;
        message.visibility = "";
        message.thumbnail = new Uint8Array(0);
        message.isDraft = false;
        if (value !== undefined)
            reflectionMergePartial<UploadVideoRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: UploadVideoRequest): UploadVideoRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string title */ 1:
                    message.title = reader.string();
                    break;
                case /* bytes file */ 2:
                    message.file = reader.bytes();
                    break;
                case /* int32 userId */ 3:
                    message.userId = reader.int32();
                    break;
                case /* bool enableComment */ 4:
                    message.enableComment = reader.bool();
                    break;
                case /* string visibility */ 5:
                    message.visibility = reader.string();
                    break;
                case /* bytes thumbnail */ 6:
                    message.thumbnail = reader.bytes();
                    break;
                case /* bool isDraft */ 7:
                    message.isDraft = reader.bool();
                    break;
                case /* google.protobuf.Timestamp postDate */ 8:
                    message.postDate = Timestamp.internalBinaryRead(reader, reader.uint32(), options, message.postDate);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: UploadVideoRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string title = 1; */
        if (message.title !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.title);
        /* bytes file = 2; */
        if (message.file.length)
            writer.tag(2, WireType.LengthDelimited).bytes(message.file);
        /* int32 userId = 3; */
        if (message.userId !== 0)
            writer.tag(3, WireType.Varint).int32(message.userId);
        /* bool enableComment = 4; */
        if (message.enableComment !== false)
            writer.tag(4, WireType.Varint).bool(message.enableComment);
        /* string visibility = 5; */
        if (message.visibility !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.visibility);
        /* bytes thumbnail = 6; */
        if (message.thumbnail.length)
            writer.tag(6, WireType.LengthDelimited).bytes(message.thumbnail);
        /* bool isDraft = 7; */
        if (message.isDraft !== false)
            writer.tag(7, WireType.Varint).bool(message.isDraft);
        /* google.protobuf.Timestamp postDate = 8; */
        if (message.postDate)
            Timestamp.internalBinaryWrite(message.postDate, writer.tag(8, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message UploadVideoRequest
 */
export const UploadVideoRequest = new UploadVideoRequest$Type();